+++ Straczek_SEW_Projekt/Mp3_Player.cs
@@ -8,10 +8,11 @@ namespace Straczek_SEW_Projekt
 {
     class Mp3_Player:Musikplayer
     {
-        public int Länge { get; set; }
-        public int Größe { get; set; }
-
-        public Mp3_Player(int id, string titel, string künstler, string album, int länge, int größe):base(id, titel, künstler, album)
+        
+        public double Länge { get; set; }
+        public double Größe { get; set; }
+         
+        public Mp3_Player(int id, string titel, string künstler, string album, double länge, double größe):base(id, titel, künstler, album)
         {
             ID = id;
             Titel = titel;
